# Implement faster join traversal