--- a/datafusion/physical-plan/src/joins/utils.rs
+++ b/datafusion/physical-plan/src/joins/utils.rs
@@ -17,7 +17,7 @@
 
 //! Join related functionality used both on logical and physical plans
 
-use std::collections::HashSet;
+use std::collections::{HashSet, VecDeque};
 use std::fmt::{self, Debug};
 use std::future::Future;
 use std::iter::once;
@@ -149,50 +149,6 @@ impl JoinHashMap {
 // Type of offsets for obtaining indices from JoinHashMap.
 pub(crate) type JoinHashMapOffset = (usize, Option<u64>);
 
-// Macro for traversing chained values with limit.
-// Early returns in case of reaching output tuples limit.
-macro_rules! chain_traverse {
-    (
-        $input_indices:ident, $match_indices:ident, $hash_values:ident, $next_chain:ident,
-        $input_idx:ident, $chain_idx:ident, $deleted_offset:ident, $remaining_output:ident
-    ) => {
-        let mut i = $chain_idx - 1;
-        loop {
-            let match_row_idx = if let Some(offset) = $deleted_offset {
-                // This arguments means that we prune the next index way before here.
-                if i < offset as u64 {
-                    // End of the list due to pruning
-                    break;
-                }
-                i - offset as u64
-            } else {
-                i
-            };
-            $match_indices.push(match_row_idx);
-            $input_indices.push($input_idx as u32);
-            $remaining_output -= 1;
-            // Follow the chain to get the next index value
-            let next = $next_chain[match_row_idx as usize];
-
-            if $remaining_output == 0 {
-                // In case current input index is the last, and no more chain values left
-                // returning None as whole input has been scanned
-                let next_offset = if $input_idx == $hash_values.len() - 1 && next == 0 {
-                    None
-                } else {
-                    Some(($input_idx, Some(next)))
-                };
-                return ($input_indices, $match_indices, next_offset);
-            }
-            if next == 0 {
-                // end of list
-                break;
-            }
-            i = next - 1;
-        }
-    };
-}
-
 // Trait defining methods that must be implemented by a hash map type to be used for joins.
 pub trait JoinHashMapType {
     /// The type of list used to store the next list
@@ -291,60 +247,45 @@ pub trait JoinHashMapType {
     fn get_matched_indices_with_limit_offset(
         &self,
         hash_values: &[u64],
-        deleted_offset: Option<usize>,
+        // TODO: HANDLE emitting in batches
         limit: usize,
         offset: JoinHashMapOffset,
     ) -> (Vec<u32>, Vec<u64>, Option<JoinHashMapOffset>) {
-        let mut input_indices = vec![];
-        let mut match_indices = vec![];
-
-        let mut remaining_output = limit;
+        let initial_capacity = hash_values.len(); // Initial capacity hint based on input size
+        let mut input_indices = Vec::with_capacity(initial_capacity);
+        let mut match_indices = Vec::with_capacity(initial_capacity);
 
         let hash_map: &RawTable<(u64, u64)> = self.get_map();
         let next_chain = self.get_list();
 
-        // Calculate initial `hash_values` index before iterating
-        let to_skip = match offset {
-            // None `initial_next_idx` indicates that `initial_idx` processing has'n been started
-            (initial_idx, None) => initial_idx,
-            // Zero `initial_next_idx` indicates that `initial_idx` has been processed during
-            // previous iteration, and it should be skipped
-            (initial_idx, Some(0)) => initial_idx + 1,
-            // Otherwise, process remaining `initial_idx` matches by traversing `next_chain`,
-            // to start with the next index
-            (initial_idx, Some(initial_next_idx)) => {
-                chain_traverse!(
-                    input_indices,
-                    match_indices,
-                    hash_values,
-                    next_chain,
-                    initial_idx,
-                    initial_next_idx,
-                    deleted_offset,
-                    remaining_output
-                );
+        // (input_index, next_offset) of remaining items to handle
+        let mut todo = VecDeque::with_capacity(initial_capacity);
 
-                initial_idx + 1
-            }
-        };
-
-        let mut row_idx = to_skip;
-        for hash_value in &hash_values[to_skip..] {
-            if let Some((_, index)) =
+        // initialize todo
+        for (input_index, hash_value) in hash_values.iter().enumerate() {
+            if let Some((_, match_row_index)) =
                 hash_map.get(*hash_value, |(hash, _)| *hash_value == *hash)
             {
-                chain_traverse!(
-                    input_indices,
-                    match_indices,
-                    hash_values,
-                    next_chain,
-                    row_idx,
-                    index,
-                    deleted_offset,
-                    remaining_output
-                );
+                input_indices.push(input_index as u32);
+                let match_row_index = match_row_index - 1;
+                match_indices.push(match_row_index);
+                let next = next_chain[match_row_index as usize];
+
+                if next != 0 {
+                    todo.push_back((input_index, next - 1));
+                }
             }
-            row_idx += 1;
+        }
+
+        // iterate the next items 
+        while let Some((input_index, match_row_idx)) = todo.pop_front() {
+            input_indices.push(input_index as u32);
+            match_indices.push(match_row_idx);
+            let next = next_chain[match_row_idx as usize];
+
+            if next != 0 {
+                todo.push_back((input_index, next - 1));
+            }
         }
 
         (input_indices, match_indices, None)