@Dandandan 2025-02-09
Update after implementing emitting in batch size: in memory performs about the same as before, but tpch_10 regressed compared to earlier implementation (so seems more having to with output batch size in some way).